+++ src/app/globals.css
@@ -152,4 +152,22 @@
     opacity: 0.9 !important;
     font-weight: 500 !important;
   }
+
+  /* Position action button below toast content */
+  .toaster [data-sonner-toast] {
+    flex-direction: column !important;
+    align-items: flex-start !important;
+  }
+  
+  .toaster [data-sonner-toast] [data-content] {
+    width: 100% !important;
+  }
+  
+  .toaster [data-sonner-toast] button[data-button],
+  .toaster [data-sonner-toast] .action-button-below {
+    margin-top: 0.75rem !important;
+    margin-left: 0 !important;
+    margin-right: auto !important;
+    align-self: flex-start !important;
+  }
 }
